pyproject.toml
[project]
name = "pandora-core-conversion"
version = "0.1.0"
description = "Pandora group conversion funnel service (loyalist -> franchisee). ADR-003."
requires-python = ">=3.12"
authors = [{ name = "Pandora Group" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.28",
    "cryptography>=43",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-python-jose>=3.3",
    "aiosqlite>=0.20",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM"]
ignore = ["B008"]  # FastAPI Depends pattern

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_return_any = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["jose.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"
